revisions before refactoring for streaming
